[TS-244] Refactor: CategoryLabel 컴포넌트, 스타일 파일을 CategoryLabel 폴더 밖으로 이동
폴더 내부에 파일이 컴포넌트, 스타일 두 개 밖에 없어서 폴더 밖으로 이동함
