+++ src/components/StarCardDetail/CategoryLabel.style.ts
